# packages/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivateResponse": {
        "properties": {
          "program": {
            "$ref": "#/components/schemas/ProgramResponse"
          },
          "scheduled_count": {
            "title": "Scheduled Count",
            "type": "integer"
          },
          "skipped_count": {
            "title": "Skipped Count",
            "type": "integer"
          }
        },
        "required": [
          "program",
          "scheduled_count",
          "skipped_count"
        ],
        "title": "ActivateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProgramDayExerciseUpdate": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "position": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "progression_strategy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProgressionStrategy"
              },
              {
                "type": "null"
              }
            ]
          },
          "rest_seconds": {
            "anyOf": [
              {
                "maximum": 3600.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rest Seconds"
          },
          "target_reps_high": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Reps High"
          },
          "target_reps_low": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Reps Low"
          },
          "target_rir_high": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Rir High"
          },
          "target_rir_low": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Rir Low"
          },
          "target_rpe_high": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Rpe High"
          },
          "target_rpe_low": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Rpe Low"
          },
          "target_sets": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Sets"
          }
        },
        "title": "ProgramDayExerciseUpdate",
        "type": "object"
      },
...
      "ProgramList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProgramListItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProgramList",
        "type": "object"
      },
...
      "ProgramResponse": {
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/ProgramDayResponse"
            },
            "title": "Days",
            "type": "array"
          },
          "days_per_week": {
            "title": "Days Per Week",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "goal": {
            "$ref": "#/components/schemas/ProgramGoal"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/ProgramSource"
          },
          "template_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "weeks": {
            "title": "Weeks",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "goal",
          "weeks",
          "days_per_week",
          "source",
          "template_id",
          "is_active",
          "activated_at",
          "days",
          "created_at"
        ],
        "title": "ProgramResponse",
        "type": "object"
      },
...
      "ProgramTemplateFull": {
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "days_per_week": {
            "title": "Days Per Week",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "goal": {
            "$ref": "#/components/schemas/ProgramGoal"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "weeks": {
            "title": "Weeks",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "author",
          "goal",
          "weeks",
          "days_per_week",
          "data"
        ],
        "title": "ProgramTemplateFull",
        "type": "object"
      },
      "ProgramTemplateList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProgramTemplateSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProgramTemplateList",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Gym API",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/program-day-exercises/{pde_id}": {
      "delete": {
        "operationId": "delete_program_day_exercise_v1_program_day_exercises__pde_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "pde_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pde Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Program Day Exercise",
        "tags": [
          "programs"
        ]
      },
      "patch": {
        "operationId": "patch_program_day_exercise_v1_program_day_exercises__pde_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "pde_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pde Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramDayExerciseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Program Day Exercise",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/program-days/{day_id}": {
      "delete": {
        "operationId": "delete_program_day_v1_program_days__day_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "day_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Day Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Program Day",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/program-days/{day_id}/exercises": {
      "post": {
        "operationId": "add_program_day_exercise_v1_program_days__day_id__exercises_post",
        "parameters": [
          {
            "in": "path",
            "name": "day_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Day Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramDayExerciseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Program Day Exercise",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/program-templates": {
      "get": {
        "operationId": "list_program_templates_v1_program_templates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramTemplateList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Program Templates",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/program-templates/{slug}": {
      "get": {
        "operationId": "get_program_template_v1_program_templates__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramTemplateFull"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Program Template",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/program-templates/{slug}/copy": {
      "post": {
        "operationId": "copy_program_template_v1_program_templates__slug__copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Copy Program Template",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/programs": {
      "get": {
        "operationId": "list_programs_v1_programs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Programs",
        "tags": [
          "programs"
        ]
      },
      "post": {
        "operationId": "create_program_v1_programs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Program",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/programs/{program_id}": {
      "delete": {
        "operationId": "delete_program_v1_programs__program_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "program_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Program Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Program",
        "tags": [
          "programs"
        ]
      },
      "get": {
        "operationId": "get_program_v1_programs__program_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "program_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Program Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Program",
        "tags": [
          "programs"
        ]
      },
      "patch": {
        "operationId": "patch_program_v1_programs__program_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "program_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Program Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Program",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/programs/{program_id}/activate": {
      "post": {
        "operationId": "activate_program_v1_programs__program_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "program_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Program Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Activate Program",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/programs/{program_id}/days": {
      "post": {
        "operationId": "create_program_day_v1_programs__program_id__days_post",
        "parameters": [
          {
            "in": "path",
            "name": "program_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Program Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramDayCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramDayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Program Day",
        "tags": [
          "programs"
        ]
      }
    },
    "/v1/programs/{program_id}/deactivate": {
      "post": {
        "operationId": "deactivate_program_v1_programs__program_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "program_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Program Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "skip_existing",
            "required": false,
            "schema": {
              "default": true,
              "title": "Skip Existing",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Deactivate Program",
        "tags": [
          "programs"
        ]
      }
    },
...
  }
}